feat: The construction process of the objects was abstracted in the builder.js file, where you can find the types of construction of the properties.
